+++ s2graphql/src/main/scala/org/apache/s2graph/graphql/bind/Unmarshaller.scala
@@ -124,52 +124,4 @@ object Unmarshaller {
       partialServiceColumns.toVector
     }
   }
-
-  def labelField(label: Label, c: Context[GraphRepository, Any]): (S2VertexLike, QueryParam) = {
-    val vertex = c.value.asInstanceOf[S2VertexLike]
-
-    val dir = c.arg[String]("direction")
-    val offset = c.arg[Int]("offset") + 1 // +1 for skip degree edge: currently not support
-    val limit = c.arg[Int]("limit")
-    val whereClauseOpt = c.argOpt[String]("filter")
-    val where = c.ctx.parser.extractWhere(label, whereClauseOpt)
-
-    val qp = QueryParam(
-      labelName = label.label,
-      direction = dir,
-      offset = offset,
-      limit = limit,
-      whereRawOpt = whereClauseOpt,
-      where = where
-    )
-
-    (vertex, qp)
-  }
-
-  def serviceColumnFieldOnService(column: ServiceColumn, c: Context[GraphRepository, Any]): (Seq[S2VertexLike], Boolean) = {
-    val ids = c.argOpt[Any]("id").toSeq ++ c.argOpt[List[Any]]("ids").toList.flatten
-    val vertices = ids.map(vid => c.ctx.toS2VertexLike(vid, column))
-
-    val columnFields = column.metasInvMap.keySet
-    val selectedFields = AstHelper.selectedFields(c.astFields)
-
-    val canSkipFetch = selectedFields.forall(f => f == "id" || !columnFields(f))
-
-    (vertices, canSkipFetch)
-  }
-
-  def serviceColumnFieldOnLabel(c: Context[GraphRepository, Any]): (S2VertexLike, Boolean) = {
-    val edge = c.value.asInstanceOf[S2EdgeLike]
-
-    val vertex = edge.tgtForVertex
-    val column = vertex.serviceColumn
-
-    val selectedFields = AstHelper.selectedFields(c.astFields)
-    val columnFields = column.metasInvMap.keySet
-
-    val canSkipFetch = selectedFields.forall(f => f == "id" || !columnFields(f))
-
-    (vertex, canSkipFetch)
-  }
-
 }
